Fork SOLA-FAO and reconfigure web services [#1 responsible:none state:fixed]
Forked SOLA-FAO for SOLA-SL-FAO. Updated namespace for EJB's and
the context root for the web services. Modified MyBatis config
for all EJBs (jdbc/sola_sl) and changed the sun-application.xml
to link to a different realm config in Glassfish (SolaSLRealm)
so the State Land services can run side by side with the SOLA
services while referencing a different database.